apps/search-service/pyproject.toml
[project]
name = "search-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api>=4.2.0",
    "fastapi>=0.124.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "uvicorn>=0.38.0",
]
